# Simplify the sum aggregators