Part of #4220

- Accept a destination type for the underlying `delta` or `cumulative`. This allows memory reuse for collections which adds a considerable optimization.
- Simplify the integration testing of the sum aggregate.
- Update benchmarking.

## Benchmarking

### Measure

```text
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric/internal/aggregate
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                                                 │   old.txt   │              new.txt               │
                                                 │   sec/op    │   sec/op     vs base               │
Sum/Int64/Cumulative/1/Measure-8                   85.97n ± 0%   89.77n ± 3%  +4.43% (p=0.000 n=10)
Sum/Int64/Cumulative/10/Measure-8                  1.076µ ± 3%   1.114µ ± 3%  +3.53% (p=0.000 n=10)
Sum/Int64/Cumulative/100/Measure-8                 11.03µ ± 3%   11.00µ ± 1%       ~ (p=0.315 n=10)
Sum/Int64/Delta/1/Measure-8                        97.38n ± 2%   97.31n ± 3%       ~ (p=0.529 n=10)
Sum/Int64/Delta/10/Measure-8                       1.127µ ± 4%   1.110µ ± 4%       ~ (p=0.271 n=10)
Sum/Int64/Delta/100/Measure-8                      11.27µ ± 4%   11.19µ ± 4%       ~ (p=0.529 n=10)
Sum/Float64/Cumulative/1/Measure-8                 97.38n ± 5%   98.12n ± 3%       ~ (p=0.529 n=10)
Sum/Float64/Cumulative/10/Measure-8                1.125µ ± 3%   1.145µ ± 4%       ~ (p=0.288 n=10)
Sum/Float64/Cumulative/100/Measure-8               11.55µ ± 2%   11.23µ ± 3%  -2.75% (p=0.002 n=10)
Sum/Float64/Delta/1/Measure-8                      98.06n ± 1%   99.47n ± 3%       ~ (p=0.165 n=10)
Sum/Float64/Delta/10/Measure-8                     1.139µ ± 3%   1.130µ ± 3%       ~ (p=0.225 n=10)
Sum/Float64/Delta/100/Measure-8                    11.31µ ± 3%   10.97µ ± 6%       ~ (p=0.060 n=10)
Sum/Precomputed/Int64/Cumulative/1/Measure-8       98.30n ± 2%   98.07n ± 2%       ~ (p=0.196 n=10)
Sum/Precomputed/Int64/Cumulative/10/Measure-8      1.122µ ± 6%   1.095µ ± 2%  -2.41% (p=0.001 n=10)
Sum/Precomputed/Int64/Cumulative/100/Measure-8     11.28µ ± 2%   10.92µ ± 6%  -3.24% (p=0.015 n=10)
Sum/Precomputed/Int64/Delta/1/Measure-8            97.87n ± 4%   97.19n ± 3%       ~ (p=0.128 n=10)
Sum/Precomputed/Int64/Delta/10/Measure-8           1.133µ ± 4%   1.098µ ± 6%  -3.09% (p=0.008 n=10)
Sum/Precomputed/Int64/Delta/100/Measure-8          11.22µ ± 3%   11.23µ ± 2%       ~ (p=0.927 n=10)
Sum/Precomputed/Float64/Cumulative/1/Measure-8     99.51n ± 3%   98.81n ± 2%       ~ (p=0.218 n=10)
Sum/Precomputed/Float64/Cumulative/10/Measure-8    1.156µ ± 4%   1.128µ ± 4%  -2.42% (p=0.050 n=10)
Sum/Precomputed/Float64/Cumulative/100/Measure-8   11.36µ ± 3%   11.10µ ± 4%       ~ (p=0.052 n=10)
Sum/Precomputed/Float64/Delta/1/Measure-8          99.93n ± 1%   98.13n ± 2%  -1.79% (p=0.019 n=10)
Sum/Precomputed/Float64/Delta/10/Measure-8         1.175µ ± 4%   1.152µ ± 3%       ~ (p=0.060 n=10)
Sum/Precomputed/Float64/Delta/100/Measure-8        11.39µ ± 2%   11.15µ ± 3%  -2.17% (p=0.015 n=10)
geomean                                            1.073µ        1.065µ       -0.77%

                                                 │   old.txt    │               new.txt               │
                                                 │     B/op     │    B/op     vs base                 │
Sum/Int64/Cumulative/1/Measure-8                   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Cumulative/10/Measure-8                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Cumulative/100/Measure-8                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/1/Measure-8                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/10/Measure-8                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/100/Measure-8                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/1/Measure-8                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/10/Measure-8                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/100/Measure-8               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/1/Measure-8                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/10/Measure-8                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/100/Measure-8                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/1/Measure-8       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/10/Measure-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/100/Measure-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/1/Measure-8            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/10/Measure-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/100/Measure-8          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/1/Measure-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/10/Measure-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/100/Measure-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/1/Measure-8          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/10/Measure-8         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/100/Measure-8        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                       ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                 │   old.txt    │               new.txt               │
                                                 │  allocs/op   │ allocs/op   vs base                 │
Sum/Int64/Cumulative/1/Measure-8                   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Cumulative/10/Measure-8                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Cumulative/100/Measure-8                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/1/Measure-8                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/10/Measure-8                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Int64/Delta/100/Measure-8                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/1/Measure-8                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/10/Measure-8                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Cumulative/100/Measure-8               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/1/Measure-8                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/10/Measure-8                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Float64/Delta/100/Measure-8                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/1/Measure-8       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/10/Measure-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Cumulative/100/Measure-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/1/Measure-8            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/10/Measure-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Int64/Delta/100/Measure-8          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/1/Measure-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/10/Measure-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Cumulative/100/Measure-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/1/Measure-8          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/10/Measure-8         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum/Precomputed/Float64/Delta/100/Measure-8        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                       ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

### ComputeAggregation

```text
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric/internal/aggregate
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                                                            │   old.txt    │               new.txt                │
                                                            │    sec/op    │    sec/op     vs base                │
Sum/Int64/Cumulative/1/ComputeAggregation-8                   235.7n ± 12%   197.2n ±  7%  -16.33% (p=0.000 n=10)
Sum/Int64/Cumulative/10/ComputeAggregation-8                  845.6n ±  3%   318.1n ±  3%  -62.39% (p=0.000 n=10)
Sum/Int64/Cumulative/100/ComputeAggregation-8                 4.869µ ±  7%   1.919µ ±  4%  -60.57% (p=0.000 n=10)
Sum/Int64/Delta/1/ComputeAggregation-8                        428.7n ±  3%   340.2n ± 13%  -20.63% (p=0.000 n=10)
Sum/Int64/Delta/10/ComputeAggregation-8                       2.067µ ±  3%   1.523µ ±  3%  -26.30% (p=0.000 n=10)
Sum/Int64/Delta/100/ComputeAggregation-8                      17.07µ ±  2%   13.55µ ±  2%  -20.62% (p=0.000 n=10)
Sum/Float64/Cumulative/1/ComputeAggregation-8                 265.2n ±  9%   212.1n ±  6%  -20.02% (p=0.000 n=10)
Sum/Float64/Cumulative/10/ComputeAggregation-8                885.6n ±  2%   332.1n ±  3%  -62.50% (p=0.000 n=10)
Sum/Float64/Cumulative/100/ComputeAggregation-8               5.136µ ±  8%   2.054µ ±  2%  -60.01% (p=0.000 n=10)
Sum/Float64/Delta/1/ComputeAggregation-8                      427.7n ±  3%   343.8n ± 11%  -19.62% (p=0.000 n=10)
Sum/Float64/Delta/10/ComputeAggregation-8                     2.092µ ±  2%   1.563µ ±  3%  -25.29% (p=0.000 n=10)
Sum/Float64/Delta/100/ComputeAggregation-8                    17.04µ ±  2%   13.66µ ±  3%  -19.84% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/1/ComputeAggregation-8       431.8n ±  6%   324.3n ± 11%  -24.89% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/10/ComputeAggregation-8      2.118µ ±  4%   1.559µ ±  4%  -26.38% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/100/ComputeAggregation-8     17.31µ ±  1%   13.61µ ±  3%  -21.40% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/1/ComputeAggregation-8            821.6n ±  4%   679.2n ±  3%  -17.34% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/10/ComputeAggregation-8           4.893µ ±  1%   4.499µ ±  3%   -8.06% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/100/ComputeAggregation-8          46.04µ ±  2%   43.98µ ±  3%   -4.48% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/1/ComputeAggregation-8     432.6n ±  4%   354.4n ± 14%  -18.06% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/10/ComputeAggregation-8    2.095µ ±  3%   1.597µ ±  3%  -23.77% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/100/ComputeAggregation-8   17.25µ ±  2%   13.61µ ±  5%  -21.07% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/1/ComputeAggregation-8          824.3n ±  3%   685.4n ±  4%  -16.84% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/10/ComputeAggregation-8         4.947µ ±  2%   4.547µ ±  3%   -8.10% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/100/ComputeAggregation-8        46.27µ ±  2%   43.60µ ±  3%   -5.77% (p=0.000 n=10)
geomean                                                       2.456µ         1.768µ        -28.03%

                                                            │    old.txt    │               new.txt                │
                                                            │     B/op      │     B/op      vs base                │
Sum/Int64/Cumulative/1/ComputeAggregation-8                     128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Int64/Cumulative/10/ComputeAggregation-8                   1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Int64/Cumulative/100/ComputeAggregation-8                  9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Int64/Delta/1/ComputeAggregation-8                          128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Int64/Delta/10/ComputeAggregation-8                        1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Int64/Delta/100/ComputeAggregation-8                       9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Float64/Cumulative/1/ComputeAggregation-8                   128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Float64/Cumulative/10/ComputeAggregation-8                 1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Float64/Cumulative/100/ComputeAggregation-8                9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Float64/Delta/1/ComputeAggregation-8                        128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Float64/Delta/10/ComputeAggregation-8                      1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Float64/Delta/100/ComputeAggregation-8                     9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/1/ComputeAggregation-8         128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/10/ComputeAggregation-8       1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/100/ComputeAggregation-8      9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/1/ComputeAggregation-8               384.0 ± 0%     288.0 ± 0%  -25.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/10/ComputeAggregation-8             1732.0 ± 0%     708.0 ± 0%  -59.12% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/100/ComputeAggregation-8          17.415Ki ± 0%   7.915Ki ± 0%  -54.55% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/1/ComputeAggregation-8       128.00 ± 0%     32.00 ± 0%  -75.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/10/ComputeAggregation-8     1056.00 ± 0%     32.00 ± 0%  -96.97% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/100/ComputeAggregation-8    9760.00 ± 0%     32.00 ± 0%  -99.67% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/1/ComputeAggregation-8             384.0 ± 0%     288.0 ± 0%  -25.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/10/ComputeAggregation-8           1732.0 ± 0%     708.0 ± 0%  -59.12% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/100/ComputeAggregation-8        17.414Ki ± 0%   7.914Ki ± 0%  -54.55% (p=0.000 n=10)
geomean                                                        1.286Ki          78.89       -94.01%

                                                            │  old.txt   │              new.txt               │
                                                            │ allocs/op  │ allocs/op   vs base                │
Sum/Int64/Cumulative/1/ComputeAggregation-8                   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Int64/Cumulative/10/ComputeAggregation-8                  2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Int64/Cumulative/100/ComputeAggregation-8                 2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Int64/Delta/1/ComputeAggregation-8                        2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Int64/Delta/10/ComputeAggregation-8                       2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Int64/Delta/100/ComputeAggregation-8                      2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Cumulative/1/ComputeAggregation-8                 2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Cumulative/10/ComputeAggregation-8                2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Cumulative/100/ComputeAggregation-8               2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Delta/1/ComputeAggregation-8                      2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Delta/10/ComputeAggregation-8                     2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Float64/Delta/100/ComputeAggregation-8                    2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/1/ComputeAggregation-8       2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/10/ComputeAggregation-8      2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Cumulative/100/ComputeAggregation-8     2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/1/ComputeAggregation-8            4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/10/ComputeAggregation-8           5.000 ± 0%   4.000 ± 0%  -20.00% (p=0.000 n=10)
Sum/Precomputed/Int64/Delta/100/ComputeAggregation-8          11.00 ± 0%   10.00 ± 0%   -9.09% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/1/ComputeAggregation-8     2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/10/ComputeAggregation-8    2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Cumulative/100/ComputeAggregation-8   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/1/ComputeAggregation-8          4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/10/ComputeAggregation-8         5.000 ± 0%   4.000 ± 0%  -20.00% (p=0.000 n=10)
Sum/Precomputed/Float64/Delta/100/ComputeAggregation-8        11.00 ± 0%   10.00 ± 0%   -9.09% (p=0.000 n=10)
geomean                                                       2.636        1.490       -43.47%
```